new task for the hypertriton 3 body becomes the default
