+++ PWGLF/NUCLEX/Hypernuclei/Hyp3Body/README.md
@@ -1,0 +1,3 @@
+The old code for the analysis of the (anti-)hyper-triton in the 3 body decay is
+in AliPhysics up to Tag vAN-20190426 
+(https://github.com/alisw/AliPhysics/releases/tag/vAN-20190426)
