+++ PWGLF/NUCLEX/PWGLFnuclexLinkDef.h
@@ -90,18 +90,12 @@
 #pragma link C++ class std::vector<SHyperTritonHe3pi>+;
 #pragma link C++ class RCollision+;
 
-
 /// * Hyp3body
-#pragma link C++ class AliAnalysisTaskHypertriton3+;
-#pragma link C++ class AliAnalysisTaskHypertriton3Dev+;
-#pragma link C++ class AliAnalysisTaskHypertriton3AOD+;
-
-/// * New Hyp3body
 #pragma link C++ class AliAnalysisTaskFindableHypertriton3+;
-#pragma link C++ class AliAnalysisTaskHypertriton3New+;
+#pragma link C++ class AliAnalysisTaskHypertriton3+;
 #pragma link C++ class AliAnalysisTaskHypertriton3ML+;
 
-/// Hypertriton Selector
+/// Hyp3body Selector
 #ifdef __CLING__
 #pragma link C++ class AliSelectorFindableHyperTriton3Body+;
 #endif
